Removed `Artax\App` functionality made obsolete by the addition of the `app.setUp` event
Prior to the addition of the built-in `app.setUp` event, the `Artax\App` boot class
included protected methods to optionally load HTTP libs based on a configuration
directive, automatically register namespaced class autoloaders and auto-require
an array of user-specified files from config.

Giving apps the ability to specify their own listeners for basic setUp obviates the
need for this functionality to be hard-wired into the framework boot class. This
also opens up the door for users to specify custom class autoloaders (if they want to)
instead of forcing the framework autoloader on them. The packaged autoloader is still
used during the boot process to register autoloading for libs in the Artax namespace.

Before: an application's config file would specify namespaces/path combinations
in key-value array pairs and have the autoloaders automatically registered.

After: the application should specify it's own event listener to register any desired
autoloaders and attach it to the `app.setUp` event.

The `Artax\Config` class and its associated test file were also updated to remove the
unnecessary references to the removed functionality.

I'm still getting used to the infinite possibilities of an event-driven approach to
application design, so we may still need to eliminate some of these sorts of things.