Introduce HttpClient interface
This allows decoupling the client entirely from the connection pool and
interceptors, which become a implementation detail of specific
implementations.

An HttpClientBuilder has been introduced, so users don't need to be
aware of the different implementations and how to compose them.

Passing a simple string instead of a Request object has been removed,
because it simplifies the public API and because there could be subtle
differences now, given that multiple implementations exist.