Do not throw when writing if socket closed
The call to shutdown() in runReadFiber will fail the request if the socket closes. The exception thrown in runWriteFiber was sometimes being caught by the event loop error handler as the top-level exception, not a previous exception as would be expected. I'm not sure how this is possible, but removing the throw fixed the issue.